requirements.txt
tqdm
dnspython